% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packing.R
\name{getContainers}
\alias{getContainers}
\title{Deprecated: Optimal packing into multiple containers}
\usage{
getContainers(profit, volume, moq, cap = 65, sold = rep(0, length(profit)))
}
\arguments{
\item{profit}{vector with profit for item}

\item{volume}{vector of item sizes in cubic meters}

\item{moq}{vector of flags where 1 means that row contans mininum order quantity (MOQ)}

\item{cap}{size of the container in cubic meters}

\item{sold}{vector with a number of items that were sold on demand}
}
\value{
vector with container numbers keeping the permutation of the original data
}
\description{
Deprecated: Optimal packing into multiple containers
}
