% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packing.R
\name{getMoqConstraint}
\alias{getMoqConstraint}
\title{MOQ contstraint generator}
\usage{
getMoqConstraint(moq)
}
\arguments{
\item{moq}{flag that indicates that this position contains MOQ}
}
\value{
matrix that expesses the MOQ constraint:
  non-MOQ item cannot be put into container that does not contain MOQ item
}
\description{
Creates matrix of moq constraints for the LP optimisation.
It is assumed that there is only one moq position per SKU and
data is sorted by sku, therefore SKU index can be calculated
}
