% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packing.R
\name{group_moq}
\alias{group_moq}
\title{Collapse function for the MOQ items}
\usage{
group_moq(units)
}
\arguments{
\item{units}{data.table with following fields: sku, utility, volume, moq}
}
\value{
data.table with sku, utility, volume and units fields. first lines for each sku are grouped according to moq
}
\description{
Combines items with MOQ greater than one to a single line that represents min amount that can be ordered
}
