% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packing.R
\name{groupFirstMoq}
\alias{groupFirstMoq}
\title{Collapse function for the MOQ items}
\usage{
groupFirstMoq(units)
}
\arguments{
\item{units}{data.table with following fields: sku, utility, volume, moq}
}
\value{
data.table with sku, utility, volume and units fields. first lines for each sku are grouped according to moq
}
\description{
Assigns earliest date to purchase order lines until first MOQ is reached within supplier-warehouse
}
