% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packing.R
\name{solveKnapsack.CBC}
\alias{solveKnapsack.CBC}
\title{Solve knapsack problem with rcbc package}
\usage{
solveKnapsack.CBC(profit, volume, moq, cap)
}
\arguments{
\item{profit}{vector with profit for item}

\item{volume}{vector of item sizes in cubic meters}

\item{moq}{vector of flags where 1 means that row contans MOQ quantity}

\item{cap}{size of the container in cubic meters}
}
\value{
vector with container numbers keeping the permutation of the original data
}
\description{
Solve knapsack problem with rcbc package
}
\seealso{
https://github.com/dirkschumacher/rcbc
}
