% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packing.R
\name{solveKnapsack}
\alias{solveKnapsack}
\title{Solves knapsack problem with the library defined
in KNAPSACK_SOLVE env variable, defaults to lpSolve package.}
\usage{
solveKnapsack(profit, volume, moq, cap)
}
\description{
Solves knapsack problem with the library defined
in KNAPSACK_SOLVE env variable, defaults to lpSolve package.
}
