% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packing.R
\name{knapsack}
\alias{knapsack}
\title{Solves knapsack problem with the library defined
in knapsack.solver option:
 - cbc (default) - uses rcbc package
 - lpsolve - uses lpSolve package}
\usage{
knapsack(profit, volume, moq = rep(0, length(profit)), cap = 65)
}
\arguments{
\item{profit}{vector with profit for item}

\item{volume}{vector of item sizes in cubic meters}

\item{moq}{vector of flags where 1 means that row contans mininum order quantity (MOQ).
Defaults to zero vector matching profit in length.}

\item{cap}{size of the container in cubic meters}
}
\value{
vector with container numbers keeping the permutation of the original data
}
\description{
Solves knapsack problem with the library defined
in knapsack.solver option:
 - cbc (default) - uses rcbc package
 - lpsolve - uses lpSolve package
}
