% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packing.R
\name{solveKnapsack}
\alias{solveKnapsack}
\title{Solves knapsack problem with the library defined
in KNAPSACK_SOLVE env variable, defaults to cbc package.}
\usage{
solveKnapsack(profit, volume, moq, cap)
}
\arguments{
\item{profit}{vector with profit for item}

\item{volume}{vector of item sizes in cubic meters}

\item{moq}{vector of flags where 1 means that row contans mininum order quantity (MOQ)}

\item{cap}{size of the container in cubic meters}
}
\value{
vector with container numbers keeping the permutation of the original data
}
\description{
Solves knapsack problem with the library defined
in KNAPSACK_SOLVE env variable, defaults to cbc package.
}
